Stop flattening arrays.
It seems the most reasonable thing for these components to accept is an
array of arrays of documents: [[{}], [{}]].
This wasn't clear initially but is now.
Also, it seems that because of none of the graph functions work with an
edge as an example, the actual input to a component needs to be an array
of arrays of vertices. That throws a monkeywrench in the works a
little, since nothing that returns an edge can be chained...