Use Faraday directly to query the Arango API
This is a pretty major refactor. The dependancy on Ashikawa was dropped
and replaced with just using Faraday to make the http requests to the
API. This means that I no longer need to work around ashikawa's habit of
stripping off Arango's internal keys or returning objects (eventually I
will want json).